requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
idna==3.7
pillow==10.3.0
python-dotenv==1.0.1
requests==2.32.2
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
